[KMTESTS:FSRTL] Add more tests for LargeMcb functions.
These are not invented but real cases: they are extracted from Ext2Fsd driver attempting to read a file from disk. The read file in ReactOS is totally corrupted.

This shows how broken our current MCB implementation is. And unfortunately confirms CORE-11002.
I recommend not to import any other FSDs relying on MCB as long as this is not fixed (eg. CORE-11040); it makes them pretty pointless.

CORE-11002


git-svn-id: https://svn.reactos.org/reactos/trunk@71091 97493ccd-5924-5043-b1f5-66cb403b36ce