[NET] Implement the "net use" usage of the net command. This allows connecting remote resources to local system with assigning them a local name, to enumerate such connected resources and to delete them.
This obsoletes the hackssign application.

The implementation is a bit... raw. It is mainly here to demonstrate what's doable in ReactOS now. And to help using features we were lacking up to now.
For instance, you can make use of 'net use * \\vmware-host\Shared Folders\{YOURSHARE}' to assign a local letter to your VMware shared folders.

ROSAPPS-303


git-svn-id: https://svn.reactos.org/reactos/trunk@72088 97493ccd-5924-5043-b1f5-66cb403b36ce